+++ activerecord/lib/active_record/type/decimal_without_scale.rb
@@ -1,0 +1,9 @@
+module ActiveRecord
+  module Type
+    class DecimalWithoutScale < ActiveModel::Type::BigInteger # :nodoc:
+      def type
+        :decimal
+      end
+    end
+  end
+end